chore(js): bump instant.page to latest
Updates [instant.page](https://instant.page/) to latest.

Was a handful of major versions behind. Considered using the unpkg.com
CDN to bypass instant.page being on the EasyPrivacy list but alas imho
it's more appropriate to just have it not work if someone using the site
has an adblocker installed. Afterall, it's just progressive enhancement.

- https://instant.page/blacklist#content-blockers